15.02.24, заполняются lemma и index таблицы v 2.

+++ src/main/java/main/model/Lemma.java
@@ -21,7 +21,7 @@ public class Lemma {
     @Column(columnDefinition = "varchar(255)", nullable = false)
     private String lemma;
     @Column(nullable = false)
-    private float frequency;
+    private int frequency;
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "site_id", nullable = false,
             foreignKey = @ForeignKey(name = "FK_lemma_site"))
